refactor: updating gatsby node file
